--- front/src/modules/timesStops/consts.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const marginRegExValidation = /^\d+(\.\d+)?%$|^\d+(\.\d+)?min\/100km$/;
